Near-complete rewrite of wrappers for Schema.
Wrappers for Key, Field, KeyBase, FieldBase, and SchemaItem are now
in the same module, and considerably more is defined in the .cc file
instead of the accompanying .py file.

This relies on a few additions to the C++, but the result should be
a much faster Python interface with a much cleaner interface.